moved ALL simulation functions into the Simulation.c source file, to make it easier to understand the program structure and which functions will need to be offloaded
