+++ src/cuda/Main.cu
@@ -29,7 +29,22 @@ int main( int argc, char* argv[] )
 	// therefore, do not profile this region!
 	// =====================================================================
 	
-	SimulationData GSD = grid_init_do_not_profile( in, mype );
+	SimulationData SD;
+
+	// If read from file mode is selected, skip initialization and load
+	// all simulation data structures from file instead
+	if( in.binary_mode == READ )
+		SD = binary_read(in);
+	else
+		SD = grid_init_do_not_profile( in, mype );
+
+	// If writing from file mode is selected, write all simulation data
+	// structures to file
+	if( in.binary_mode == WRITE && mype == 0 )
+		binary_write(in, SD);
+
+	// Move data to GPU
+	SimulationData GSD = move_simulation_data_to_device( in, mype, SD );
 
 	// =====================================================================
 	// Cross Section (XS) Parallel Lookup Simulation
@@ -54,7 +69,7 @@ int main( int argc, char* argv[] )
 		verification = run_event_based_simulation(in, GSD, mype);
 	else
 	{
-		printf("History-based simulation not implemented in CUDA code. Use event-based method with \"-m event\" argument\n");
+		printf("History-based simulation not implemented in CUDA code. Instead,\nuse the event-based method with \"-m event\" argument.\n");
 		exit(1);
 	}
