--- src/openmp-threading/CalculateXS.c
@@ -1,205 +1,0 @@
-#include "XSbench_header.h"
-
-// Calculates the microscopic cross section for a given nuclide & energy
-void calculate_micro_xs(   double p_energy, int nuc, long n_isotopes,
-                           long n_gridpoints,
-                           double * restrict egrid, int * restrict index_data,
-                           NuclideGridPoint * restrict nuclide_grids,
-                           long idx, double * restrict xs_vector, int grid_type, int hash_bins ){
-	// Variables
-	double f;
-	NuclideGridPoint * low, * high;
-
-	// If using only the nuclide grid, we must perform a binary search
-	// to find the energy location in this particular nuclide's grid.
-	if( grid_type == NUCLIDE )
-	{
-		// Perform binary search on the Nuclide Grid to find the index
-		idx = grid_search_nuclide( n_gridpoints, p_energy, &nuclide_grids[nuc*n_gridpoints], 0, n_gridpoints-1);
-
-		// pull ptr from nuclide grid and check to ensure that
-		// we're not reading off the end of the nuclide's grid
-		if( idx == n_gridpoints - 1 )
-			low = &nuclide_grids[nuc*n_gridpoints + idx - 1];
-		else
-			low = &nuclide_grids[nuc*n_gridpoints + idx];
-	}
-	else if( grid_type == UNIONIZED) // Unionized Energy Grid - we already know the index, no binary search needed.
-	{
-		// pull ptr from energy grid and check to ensure that
-		// we're not reading off the end of the nuclide's grid
-		if( index_data[idx * n_isotopes + nuc] == n_gridpoints - 1 )
-			low = &nuclide_grids[nuc*n_gridpoints + index_data[idx * n_isotopes + nuc] - 1];
-		else
-			low = &nuclide_grids[nuc*n_gridpoints + index_data[idx * n_isotopes + nuc]];
-	}
-	else // Hash grid
-	{
-		// load lower bounding index
-		int u_low = index_data[idx * n_isotopes + nuc];
-
-		// Determine higher bounding index
-		int u_high;
-		if( idx == hash_bins - 1 )
-			u_high = n_gridpoints - 1;
-		else
-			u_high = index_data[(idx+1)*n_isotopes + nuc] + 1;
-
-		// Check edge cases to make sure energy is actually between these
-		// Then, if things look good, search for gridpoint in the nuclide grid
-		// within the lower and higher limits we've calculated.
-		double e_low  = nuclide_grids[nuc*n_gridpoints + u_low].energy;
-		double e_high = nuclide_grids[nuc*n_gridpoints + u_high].energy;
-		int lower;
-		if( p_energy <= e_low )
-			lower = 0;
-		else if( p_energy >= e_high )
-			lower = n_gridpoints - 1;
-		else
-			lower = grid_search_nuclide( n_gridpoints, p_energy, &nuclide_grids[nuc*n_gridpoints], u_low, u_high);
-
-		if( lower == n_gridpoints - 1 )
-			low = &nuclide_grids[nuc*n_gridpoints + lower - 1];
-		else
-			low = &nuclide_grids[nuc*n_gridpoints + lower];
-	}
-	
-	high = low + 1;
-	
-	// calculate the re-useable interpolation factor
-	f = (high->energy - p_energy) / (high->energy - low->energy);
-
-	// Total XS
-	xs_vector[0] = high->total_xs - f * (high->total_xs - low->total_xs);
-	
-	// Elastic XS
-	xs_vector[1] = high->elastic_xs - f * (high->elastic_xs - low->elastic_xs);
-	
-	// Absorbtion XS
-	xs_vector[2] = high->absorbtion_xs - f * (high->absorbtion_xs - low->absorbtion_xs);
-	
-	// Fission XS
-	xs_vector[3] = high->fission_xs - f * (high->fission_xs - low->fission_xs);
-	
-	// Nu Fission XS
-	xs_vector[4] = high->nu_fission_xs - f * (high->nu_fission_xs - low->nu_fission_xs);
-	
-	//test
-	/*	
-	if( omp_get_thread_num() == 0 )
-	{
-		printf("Lookup: Energy = %lf, nuc = %d\n", p_energy, nuc);
-		printf("e_h = %lf e_l = %lf\n", high->energy , low->energy);
-		printf("xs_h = %lf xs_l = %lf\n", high->elastic_xs, low->elastic_xs);
-		printf("total_xs = %lf\n\n", xs_vector[1]);
-	}
-	*/
-	
-}
-
-// Calculates macroscopic cross section based on a given material & energy 
-void calculate_macro_xs( double p_energy, int mat, long n_isotopes,
-                         long n_gridpoints, int * restrict num_nucs,
-                         double * restrict concs,
-                         double * restrict egrid, int * restrict index_data,
-                         NuclideGridPoint * restrict nuclide_grids,
-                         int * restrict mats,
-                         double * restrict macro_xs_vector, int grid_type, int hash_bins, int max_num_nucs ){
-	int p_nuc; // the nuclide we are looking up
-	long idx = -1;	
-	double conc; // the concentration of the nuclide in the material
-
-	// cleans out macro_xs_vector
-	for( int k = 0; k < 5; k++ )
-		macro_xs_vector[k] = 0;
-
-	// If we are using the unionized energy grid (UEG), we only
-	// need to perform 1 binary search per macroscopic lookup.
-	// If we are using the nuclide grid search, it will have to be
-	// done inside of the "calculate_micro_xs" function for each different
-	// nuclide in the material.
-	if( grid_type == UNIONIZED )
-		idx = grid_search( n_isotopes * n_gridpoints, p_energy, egrid);	
-	else if( grid_type == HASH )
-	{
-		double du = 1.0 / hash_bins;
-		idx = p_energy / du;
-	}
-	
-	// Once we find the pointer array on the UEG, we can pull the data
-	// from the respective nuclide grids, as well as the nuclide
-	// concentration data for the material
-	// Each nuclide from the material needs to have its micro-XS array
-	// looked up & interpolatied (via calculate_micro_xs). Then, the
-	// micro XS is multiplied by the concentration of that nuclide
-	// in the material, and added to the total macro XS array.
-	// (Independent -- though if parallelizing, must use atomic operations
-	//  or otherwise control access to the xs_vector and macro_xs_vector to
-	//  avoid simulataneous writing to the same data structure)
-	for( int j = 0; j < num_nucs[mat]; j++ )
-	{
-		double xs_vector[5];
-		p_nuc = mats[mat*max_num_nucs + j];
-		conc = concs[mat*max_num_nucs + j];
-		calculate_micro_xs( p_energy, p_nuc, n_isotopes,
-		                    n_gridpoints, egrid, index_data,
-		                    nuclide_grids, idx, xs_vector, grid_type, hash_bins );
-		for( int k = 0; k < 5; k++ )
-			macro_xs_vector[k] += xs_vector[k] * conc;
-	}
-	
-	//test
-	/*
-	for( int k = 0; k < 5; k++ )
-		printf("Energy: %lf, Material: %d, XSVector[%d]: %lf\n",
-		       p_energy, mat, k, macro_xs_vector[k]);
-	*/
-}
-
-
-// (fixed) binary search for energy on unionized energy grid
-// returns lower index
-long grid_search( long n, double quarry, double * restrict A)
-{
-	long lowerLimit = 0;
-	long upperLimit = n-1;
-	long examinationPoint;
-	long length = upperLimit - lowerLimit;
-
-	while( length > 1 )
-	{
-		examinationPoint = lowerLimit + ( length / 2 );
-		
-		if( A[examinationPoint] > quarry )
-			upperLimit = examinationPoint;
-		else
-			lowerLimit = examinationPoint;
-		
-		length = upperLimit - lowerLimit;
-	}
-	
-	return lowerLimit;
-}
-
-// binary search for energy on nuclide energy grid
-long grid_search_nuclide( long n, double quarry, NuclideGridPoint * A, long low, long high)
-{
-	long lowerLimit = low;
-	long upperLimit = high;
-	long examinationPoint;
-	long length = upperLimit - lowerLimit;
-
-	while( length > 1 )
-	{
-		examinationPoint = lowerLimit + ( length / 2 );
-		
-		if( A[examinationPoint].energy > quarry )
-			upperLimit = examinationPoint;
-		else
-			lowerLimit = examinationPoint;
-		
-		length = upperLimit - lowerLimit;
-	}
-	
-	return lowerLimit;
-}
