+++ src/openmp-threading/XSutils.c
@@ -26,20 +26,6 @@ int NGP_compare(const void * a, const void * b)
 		return 0;
 }
 
-// Park & Miller Multiplicative Conguential Algorithm
-// From "Numerical Recipes" Second Edition
-double rn(unsigned long * seed)
-{
-	double ret;
-	unsigned long n1;
-	unsigned long a = 16807;
-	unsigned long m = 2147483647;
-	n1 = ( a * (*seed) ) % m;
-	*seed = n1;
-	ret = (double) n1 / m;
-	return ret;
-}
-
 // RNG Used for Verification Option.
 // This one has a static seed (must be set manually in source).
 // Park & Miller Multiplicative Conguential Algorithm
